frontend: App: Extract defaultAppThemes to its own file
Because importing the themes from themeSlice causes a dependency
on all of redux.
